rework nanospinner implementation to not require printing inbetween
the previous implementation was so messy that I literally thought about 
how to do this better during a dream last night.

New convention is that only the final success log of the script itself 
is using cyan color. For errors spinner.error() is called with no 
argument followed by die().

Functions such as the quest descriptor now return a log object instead 
of printing it themself.